requirements.txt
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.3
boto3==1.34.69
botocore==1.34.69
cachetools==5.3.3
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.7
crispy-bootstrap5==2023.10
cryptography==39.0.1
defusedxml==0.7.1
distro==1.9.0
dj-database-url==2.1.0
Django==5.0.2
django-allauth==0.61.1
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-debug-toolbar==4.2.0
django-recaptcha==4.0.0
django-storages==1.14.2
fastapi==0.110.0
frozenlist==1.4.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.4
Jinja2==3.1.3
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marvin==2.3.0
mdurl==0.1.2
multidict==6.0.5
oauthlib==3.2.2
openai==1.14.2
packaging==23.1
partialjson==0.0.5
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
Pygments==2.17.2
PyJWT==2.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
regex==2023.12.25
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.7.1
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.4.3
starlette==0.36.3
tiktoken==0.6.0
tqdm==4.66.2
typer==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==1.26.14
uvicorn==0.29.0
whitenoise==6.6.0
yarl==1.9.4
